# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strong Hire API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/magic-link": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Magic Link",
        "operationId": "send_magic_link_auth_magic_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkSent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/magic-link/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Magic Link Callback",
        "operationId": "magic_link_callback_auth_magic_link_callback_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Signup",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/dev-login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Dev Login",
        "description": "Local only (APP_ENV=local). Signs in any email with no check.",
        "operationId": "dev_login_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthProvider": {
        "type": "string",
        "enum": [
          "google",
          "email",
          "dev"
        ],
        "title": "AuthProvider"
      },
...
      "AuthState": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "signed_out",
              "needs_signup",
              "signed_in"
            ],
            "title": "Status"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Set for needs_signup and signed_in."
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthUser"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AuthState"
      },
      "AuthUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "auth_provider": {
            "$ref": "#/components/schemas/AuthProvider"
          },
          "training_consent": {
            "type": "boolean",
            "title": "Training Consent"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "description": "True for ADMIN_EMAILS users: the web app shows Admin.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "email",
          "auth_provider",
          "training_consent",
          "created_at"
        ],
        "title": "AuthUser"
      },
...
      "DevLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "title": "Email",
            "default": "dev@example.com"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DevLoginRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MagicLinkRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email"
        ],
        "title": "MagicLinkRequest"
      },
      "MagicLinkSent": {
        "properties": {
          "sent": {
            "type": "boolean",
            "title": "Sent"
          },
          "dev_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Link",
            "description": "Only when APP_ENV=local, so no email server is needed."
          }
        },
        "type": "object",
        "required": [
          "sent"
        ],
        "title": "MagicLinkSent"
      },
...
      "SignupRequest": {
        "properties": {
          "age_confirmed": {
            "type": "boolean",
            "title": "Age Confirmed",
            "description": "The user confirmed they are 18 or older."
          },
          "terms_accepted": {
            "type": "boolean",
            "title": "Terms Accepted"
          },
          "training_consent": {
            "type": "boolean",
            "title": "Training Consent",
            "description": "AC-2: off unless the user opts in.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "age_confirmed",
          "terms_accepted"
        ],
        "title": "SignupRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}